dev: restructure project structure
