+++ qrcode/__init__.py
@@ -1,0 +1,2 @@
+from .creator import QRCodeCreatorMenu
+from .reader import QRCodeReaderMenu
