+++ utils/__init__.py
@@ -1,0 +1,1 @@
+from .callbacks import no_callback
