+++ utils/callbacks.py
@@ -1,0 +1,2 @@
+def no_callback(**kwargs):
+    pass
